+++ src/lib/utils/notionClient.ts
@@ -1,0 +1,7 @@
+const { Client } = require("@notionhq/client")
+
+const token = process.env.NOTION_TOKEN
+
+export const notion = new Client({
+  auth: token,
+})
